Be/feat30 기본 API 구현 - JDBC template 적용 (#107)
* [fix] car domain Set -> List 로직 Repository에서 분리

- CarMapper에서 분리 진행
- final 등 코드 컨벤션 정비

[issue] #30

* [fix] Model, Trim 코드 컨벤션 정리

[issue] #30

* [refactor] car domain

- Service, Controller에서 interface 도입
- 각 역할을 더 잘 표현하기 위해 dto, method 이름 수정

[issue] #30

* [refactor] trim domain

- Service, Controller에 interface 도입
- 각 역할을 더 잘 표현하기 위해 dto, method 이름 수정

[issue] #30

* [refactor] model domain

- Service, Controller에 interface 도입
- 각 역할을 더 잘 표현하기 위해 dto, method 이름 수정

[issue] #30

* [fix] Spring Data JDBC에서 JDBC template로 전환

- 성능 문제로 Spring Data JDBC 기능을 사용하지 않고 native query를 사용함에 따라 Spring Data JDBC을 유지할 필요성을 느끼지 못해, JDBC template으로 전환

[issue] #30

* [fix] car domain JDBC template 적용

- CarRepositoryDefaultImpl 구현
- MappedCollection 삭제

[issue] #30

* [fix] color domain JDBC template 적용

- 관계 테이블의 entity class에 id 추가
- MappedCollection 삭제

[issue] #30

* [fix] model domain JDBC template 적용

- ModelRepositoryDefaultImpl 구현
- MappedCollection 삭제
- 연관 관계 테이블 entity class에 id 추가
- ModelOptionsDTO의 특성을 잘 나타내기 위해 ModelOptionsSortedDTO로 이름 변경

[issue] #30

* [fix] car domain 코드 컨벤션 정리

- interface에서 annotation 삭제
- 구현 클래스 메소드에 @Override 추가

[issue] #30

* [fix] option domain JDBC template 적용

- entity class에 관계 테이블 id 추가
- MappedCollection 삭제

[issue] #30

* [fix] tag domain JDBC template 적용

- entity class에 관계 테이블 id 추가
- MappedCollection 삭제

[issue] #30

* [fix] trim domain JDBC template 적용

- entity class에 관계 테이블 id 추가
- MappedCollection 삭제
- interface에서 annotation 삭제

[issue] #30

* [fix] ModelRepositoryDefaultImpl 에러 해결

- n개의 변수를 사용해야 하는 상황
- NamedParameterJdbcTemplate을 사용해서 해결

[issue] #30
