+++ BackEnd/FourEver/src/main/java/A4/FourEver/domain/car/domain/Car.java
@@ -1,12 +1,10 @@
 package A4.FourEver.domain.car.domain;
 
 import lombok.Getter;
-import org.springframework.data.annotation.Id;
 
 @Getter
 public class Car {
-
-    @Id
     private Long id;
     private String name;
+    private Long car_category_id;
 }
