+++ BackEnd/FourEver/src/main/java/A4/FourEver/domain/model/api/ModelController.java
@@ -1,30 +1,7 @@
 package A4.FourEver.domain.model.api;
 
-import A4.FourEver.domain.model.application.ModelService;
-import A4.FourEver.domain.model.dto.ModelOptionConfigDTO;
-import io.swagger.v3.oas.annotations.Operation;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import A4.FourEver.domain.model.dto.ModelOptionsSortedDTO;
 
-@RestController
-@RequestMapping("/model")
-public class ModelController {
-
-    private final ModelService modelService;
-
-    public ModelController(ModelService modelService) {
-        this.modelService = modelService;
-    }
-
-    @Operation(summary = "해당 모델의 옵션 정보 조회")
-    @GetMapping("/option")
-    public ModelOptionConfigDTO getModelOption(
-            @RequestParam("trim_id") final Long trim_id,
-            @RequestParam("body_id") final Long body_id,
-            @RequestParam("engine_id") final Long engine_id,
-            @RequestParam("drive_id") final Long drive_id) {
-        return modelService.getModelOption(trim_id, body_id, engine_id, drive_id);
-    }
+public interface ModelController {
+    ModelOptionsSortedDTO getModelOption(final Long trim_id, final Long body_id, final Long engine_id, final Long drive_id);
 }
