+++ BackEnd/FourEver/src/main/java/A4/FourEver/domain/option/defaultOption/domain/DefaultOption.java
@@ -1,14 +1,12 @@
 package A4.FourEver.domain.option.defaultOption.domain;
 
 import lombok.Getter;
-import org.springframework.data.annotation.Id;
 
 @Getter
 public class DefaultOption {
-
-    @Id
     private Long id;
     private String name;
     private String description;
     private String image;
+    private Long default_option_category_id;
 }
