+++ BackEnd/FourEver/src/main/java/A4/FourEver/domain/tag/exColorTag/domain/ExColorTag.java
@@ -1,13 +1,11 @@
 package A4.FourEver.domain.tag.exColorTag.domain;
 
 import lombok.Getter;
-import org.springframework.data.annotation.Id;
 
 @Getter
 public class ExColorTag {
-
-    @Id
     private Long id;
     private String name;
     private Long count;
+    private Long ex_color_id;
 }
