+++ BackEnd/FourEver/src/main/java/A4/FourEver/domain/tag/extraOptionTag/domain/ExtraOptionTag.java
@@ -1,13 +1,11 @@
 package A4.FourEver.domain.tag.extraOptionTag.domain;
 
 import lombok.Getter;
-import org.springframework.data.annotation.Id;
 
 @Getter
 public class ExtraOptionTag {
-
-    @Id
     private Long id;
     private String name;
     private Long count;
+    private Long extra_option_id;
 }
