+++ BackEnd/FourEver/src/main/java/A4/FourEver/domain/tag/inColorTag/domain/InColorTag.java
@@ -1,13 +1,11 @@
 package A4.FourEver.domain.tag.inColorTag.domain;
 
 import lombok.Getter;
-import org.springframework.data.annotation.Id;
 
 @Getter
 public class InColorTag {
-
-    @Id
     private Long id;
     private String name;
     private Long count;
+    private Long in_color_id;
 }
