+++ BackEnd/FourEver/src/main/java/A4/FourEver/domain/trim/drive/domain/Drive.java
@@ -1,15 +1,13 @@
 package A4.FourEver.domain.trim.drive.domain;
 
 import lombok.Getter;
-import org.springframework.data.annotation.Id;
 
 @Getter
 public class Drive {
-
-    @Id
     private Long id;
     private String name;
     private String image;
     private String description;
     private Double price;
+    private Long car_id;
 }
