+++ BackEnd/FourEver/src/main/java/A4/FourEver/domain/trim/engine/domain/Engine.java
@@ -1,12 +1,9 @@
 package A4.FourEver.domain.trim.engine.domain;
 
 import lombok.Getter;
-import org.springframework.data.annotation.Id;
 
 @Getter
 public class Engine {
-
-    @Id
     private Long id;
     private String name;
     private String image;
@@ -14,5 +11,5 @@ public class Engine {
     private String max_output;
     private String max_tok;
     private Double price;
+    private Long car_id;
 }
-
